requirements.txt
Django==5.1.5
djangorestframework==3.14.0
tensorflow==2.12.0
transformers==4.30.0
spacy==3.6.0
requests==2.28.1

#for notebooks
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.6.1
torch== 2.6.0

#file does not include the spaCy language model (en_core_web_sm). After installing the dependencies, run on console: python -m spacy download en_core_web_sm